* Use consistently float and double types * Fix all compilation warnings

+++ src/Curve_selection.h
@@ -6,7 +6,7 @@ struct Curve_selection
     // The curve selection is normalized to [0, 1]
     float t_start, t_end;
 
-    bool in_range(double t) const
+    bool in_range(float t) const
     {
         return (t_start <= t && t <= t_end);
     }
